requirements.txt
Flask==2.3.2
openai==1.8.4
python-dotenv==1.0.0
